Refactor express application into own function.
Moved express server from 'server.js' into its own file 'app.js' so that
it can be tested.
