+++ back/server.js
@@ -1,28 +1,9 @@
 /**** Express imports ****/
-const express = require('express');
-const app = express();
-const bodyParser = require('body-parser');
-const cookieParser = require('cookie-parser');
-const helmet = require('helmet');
-const router = require('./routes');
+const app = require('./app.js')();
 const os = require('os');
-const morgan = require('morgan');
 
 require('dotenv').config();
 const port = process.env.ALT_PORT || process.env.PORT || 8000; //azure gives port as an environment variable
-
-app.use(helmet());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(bodyParser.json());
-app.use(cookieParser());
-
-//server calls are all in the routes
-app.use('/api', morgan('short'));
-app.use('/api', router);
-
-
-// Rendering the front-end react app
-app.use('/', express.static('../front/build/'));
 
 const server = app.listen(port, () => {
   console.error('Server on ' + port);
